[Relay][VM] Allow to config allocator type and refactor vm code structure
