+++ src/relay/backend/vm/removed_unused_funcs.cc
@@ -26,7 +26,6 @@
 #include <tvm/relay/expr.h>
 #include <tvm/relay/expr_functor.h>
 #include <tvm/relay/transform.h>
-#include <tvm/runtime/vm.h>
 #include <tvm/support/logging.h>
 
 #include <iostream>
